requirements.txt
pydantic~=2.9.2
joblib~=1.5.1
fastapi~=0.115.4
requests~=2.32.3
sentence-transformers~=5.0.0
scikit-learn~=1.7.1
pandas~=2.3.1
python-dotenv~=1.0.1
httpx~=0.27.2
numpy~=2.3.2
faiss-cpu~=1.12.0
